--- app/(root)/about/layout.js
@@ -1,11 +1,0 @@
-import LayoutProvider from "./layoutProvider";
-
-
-export const metadata = {
-  title: "About us",
-};
-
-export default function layout({ children }) {
-  return <LayoutProvider>{children}</LayoutProvider>;
-
-}
