--- app/(root)/mandate/layout.js
@@ -1,11 +1,0 @@
-import LayoutProvider from "./layoutProvider";
-
-
-export const metadata = {
-  title: "Mandate",
-};
-
-export default function layout({ children }) {
-  return <LayoutProvider>{children}</LayoutProvider>;
-
-}
