+++ app/(root)/newsandevents/layout.tsx
@@ -1,0 +1,20 @@
+import LayoutProvider from "./layoutProvider";
+
+
+
+export const metadata = {
+  title: "News & Events",
+};
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+
+    <LayoutProvider>
+      {children}
+    </LayoutProvider>
+  );
+}
