+++ app/(root)/newsandevents/layoutProvider.tsx
@@ -6,23 +6,19 @@ import { usePathname } from "next/navigation";
 
 const Tabs = [
   {
-    name: "Research",
+    name: "News",
     status: "active",
-    url: "/mandate",
+    url: "/newsandevents",
   },
   {
-    name: "Training",
+    name: "Events",
     status: "active",
-    url: "/mandate/training",
+    url: "/newsandevents/events",
   },
-  {
-    name: "Extension",
-    status: "active",
-    url: "/mandate/initiatives",
-  },
+
 ];
 
-export default function LayoutProvider({ children }) {
+ const LayoutProvider = ({ children }: {children: React.ReactNode}) => {
   const currentRoute = usePathname();
 
   return (
@@ -31,7 +27,7 @@ export default function LayoutProvider({ children }) {
 
         <div className="w-full">
           <div className="w-full whitespace-nowrap">
-            <ul className="flex text-sm justify-center text-center lg:text-lg gap-4 border-b -mb-px">
+            <ul className="flex text-sm justify-evenly lg:text-lg gap-4 border-b -mb-px">
               {Tabs.map(({ name, status, url }, index) => {
                 const activeTab = cn(currentRoute === url ? "font-bold font-Satoshi_Bold border-b-2" : "border-b-0 text-[#667085] font-Satoshi_Medium", "inline-block py-4 px-3.5 text-sm");
 
@@ -50,3 +46,5 @@ export default function LayoutProvider({ children }) {
     </div>
   );
 }
+
+export default LayoutProvider;
