--- app/(root)/programsandprojects/layout.ts
@@ -1,11 +1,0 @@
-import LayoutProvider from "./layoutProvider";
-
-
-export const metadata = {
-  title: "Programs & Projects",
-};
-
-export default function layout({ children }) {
-  return <LayoutProvider>{children}</LayoutProvider>;
-
-}
