+++ app/(root)/programsandprojects/layout.tsx
@@ -1,0 +1,19 @@
+import LayoutProvider from "./layoutProvider";
+
+
+export const metadata = {
+  title: "Programs & Projects",
+};
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+
+    <LayoutProvider>
+      {children}
+    </LayoutProvider>
+  );
+}
