+++ components/About/Mission/Value.tsx
