--- components/Mandate/Mandate/Research.tsx
@@ -1,9 +1,0 @@
-const Research = () => {
-    return (
-        <main>
-            
-        </main>
-    )
-}
-
-export default Research;
